% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_comparison.R
\name{density_comparison}
\alias{density_comparison}
\title{Density Comparison}
\usage{
density_comparison(d, density_threshold = 1)
}
\arguments{
\item{d}{Density of liquid (g/mL).}

\item{density}{Average density of water (threshold). Default is 1 g/mL.}
}
\value{
Response of whether average density of sample liquid is more, less, or equal to the average density of water.
}
\description{
Compares mean density of samples to density of water.
}
