% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_sample_data.R
\docType{data}
\name{density_sample_data}
\alias{density_sample_data}
\title{Density Sample Data}
\format{
A data frame with 50 rows and 1 column. Mean = 1 and SD = 0.5.
\describe{
\item{d}{Liquid densities (g/mL)sldjkflskdjfksdjf;sdf;lsd;mcd;sm;ksdjflksnfdklsndklfnsklfnsklfnsklnfsdlkf}
\item{yes}{Yes}
}
}
\usage{
density_sample_data
}
\description{
Sample normal data for densities of liquids.
}
\keyword{datasets}
