% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_fun.R
\name{fish_fun}
\alias{fish_fun}
\title{Fish Species Summary}
\usage{
fish_fun(species, common, rare, total, hist = FALSE)
}
\arguments{
\item{species}{Factor of fish names.}

\item{common}{Most common fish.}

\item{rare}{Rarest fish.}

\item{total}{Total number of fish.}

\item{hist}{Whether to return a histogram with a title of the total number of fish caught. Default is FALSE.}
}
\value{
List of common, rare, and total species. Will also return a histogram if hist = TRUE.
}
\description{
Returns the most common species, the rarest species, and the total number of fish from a factor of fish names. Optional to also return a histogram.
}
